requirements.txt
sentence-transformers
torch
ollama
pillow